Make tests windows only. Adjust assertions for local vs azure db.
